src/deepearth/encoders/language/requirements.txt
# Core dependencies for DeepSeek Language Encoder
flask>=2.3.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
llama-cpp-python>=0.2.0

# UMAP and clustering
umap-learn>=0.5.0
hdbscan>=0.8.0
scikit-learn>=1.3.0

# Optional: For better HSL color generation
# hsluv>=5.0.0

# Google Cloud (optional, for deployment)
# google-cloud-storage>=2.10.0